requirements.txt
Django==5.0.6
python-dotenv==1.0.0
mysqlclient==2.2.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
requests==2.31.0